refactor: getrusage wrapper to common
De-duplicate the wrapper to measure CPU usage via getrusage(), we are
going to need a third copy for Pub/Sub.
