+++ google/cloud/storage/benchmarks/storage_throughput_vs_cpu_benchmark.cc
@@ -287,7 +287,7 @@ TestResults RunThread(ThroughputOptions const& options,
 
   auto deadline = std::chrono::steady_clock::now() + options.duration;
 
-  gcs_bm::SimpleTimer timer;
+  gcs_bm::CpuUsage timer;
   TestResults results;
 
   std::int32_t iteration_count = 0;
