+++ google/cloud/storage/benchmarks/throughput_experiment.cc
@@ -60,7 +60,7 @@ class UploadObject : public ThroughputExperiment {
     // ObjectInsert()
     if (static_cast<std::size_t>(config.object_size) < random_data_.size() &&
         prefer_insert_) {
-      SimpleTimer timer;
+      CpuUsage timer;
       timer.Start();
       std::string data =
           random_data_.substr(0, static_cast<std::size_t>(config.object_size));
@@ -80,7 +80,7 @@ class UploadObject : public ThroughputExperiment {
                               timer.cpu_time(),
                               object_metadata.status()};
     }
-    SimpleTimer timer;
+    CpuUsage timer;
     timer.Start();
     auto writer = client_.WriteObject(
         bucket_name, object_name,
@@ -137,7 +137,7 @@ class DownloadObject : public ThroughputExperiment {
 
     std::vector<char> buffer(config.app_buffer_size);
 
-    SimpleTimer timer;
+    CpuUsage timer;
     timer.Start();
     auto reader = client_.ReadObject(
         bucket_name, object_name,
@@ -189,7 +189,7 @@ class DownloadObjectLibcurl : public ThroughputExperiment {
     auto header = creds_->AuthorizationHeader();
     if (!header) return {};
 
-    SimpleTimer timer;
+    CpuUsage timer;
     timer.Start();
     struct curl_slist* slist1 = nullptr;
     slist1 = curl_slist_append(slist1, header->c_str());
@@ -265,7 +265,7 @@ class DownloadObjectRawGrpc : public ThroughputExperiment {
   ThroughputResult Run(std::string const& bucket_name,
                        std::string const& object_name,
                        ThroughputExperimentConfig const& config) override {
-    SimpleTimer timer;
+    CpuUsage timer;
 
     timer.Start();
     google::storage::v1::GetObjectMediaRequest request;
