+++ google/cloud/testing_util/google_cloud_cpp_testing.bzl
@@ -23,6 +23,7 @@ google_cloud_cpp_testing_hdrs = [
     "chrono_literals.h",
     "command_line_parsing.h",
     "contains_once.h",
+    "cpu_usage.h",
     "crash_handler.h",
     "example_driver.h",
     "expect_exception.h",
@@ -37,6 +38,7 @@ google_cloud_cpp_testing_srcs = [
     "assert_ok.cc",
     "capture_log_lines_backend.cc",
     "command_line_parsing.cc",
+    "cpu_usage.cc",
     "crash_handler.cc",
     "example_driver.cc",
     "scoped_environment.cc",
